+++ MPC.cpp
@@ -7,9 +7,11 @@
 #include <iostream>
 #include <assert.h>
 #include "picosha2.h"
+
+ 
 using namespace dpf;
 
-typedef uint8_t leaf_t;
+typedef uint16_t leaf_t;
 typedef __m256i node_t;
 typedef LowMC<__m256i> prgkey_t;
 typedef block<node_t> block_t;
@@ -39,239 +41,233 @@ inline bool xor_if(const bool & block1, const bool & block2, bool flag)
   }  
 }
 
-class MPCrandomness
-{
-  
-  public:
-  
-    MPCrandomness(AES_KEY& aeskey, __m128i seed, size_t len) :
-    buf((unsigned char *)malloc(len * sizeof(seed))),
-    cur(buf)
-    {
-      PRG(aeskey, seed, (__m128i *)buf, len);
-    } 
-    ~MPCrandomness()
-    {
-    free(buf);
-    }
-  
 
 
 
-  inline bool next(int & val)
-  {
-    memcpy(&val, cur, sizeof(int));
-    val &= 1;
-    cur += sizeof(int);
-    bool val_b = val;
-
-    //std::cout << "val_b = " << val_b << std::endl;
-
-    return val_b;
-  }
-
-  inline bool next_bool()
-  {
-    int  val;
-    memcpy(&val, cur, sizeof(int));
-    val &= 1;
-    cur += sizeof(int);
-    bool val_b = val;
-
-    //std::cout << "val_b = " << val_b << std::endl;
-
-    return val_b;
-  }
-
-  template<typename T>
-  inline T& next(T & val)
-  {
-    memcpy(&val, cur, sizeof(T));
-    cur += sizeof(T);
-    return val;
-  }
-  
-  inline block_t next_block()
-  {
-    block_t val;
-    memcpy(&val, cur, sizeof(block_t));
-    cur += sizeof(block_t);
-    return val;
-  }
-  
-
-  template<typename T>
-  inline T && next()
-  {
-    T val;
-    memcpy(&val, cur, sizeof(T));
-    cur += sizeof(T);
-    return val;
-  }
-
-
-  private:
-  unsigned char * buf;
-  unsigned char * cur;
-};
-
-
-
+#include "randomness.h"
 #include "transcripts.h"
 #include "simulator.h"
 #include "verifier.h"
+#include "verifier2.h"
 
+// size_t number_of_simulations = 10;
+
+ 
 int main(int argc, char * argv[])
-{ 
-
-
+{  
   std::cout << "rounds = " << rounds << std::endl;
 
   bool party0 = false; bool party1 = true;
   
 
 
-  const size_t nitems = 1ULL << 44;
-  const size_t target = atoi(argv[1]);
+  const size_t nitems = 1ULL << 40;
+  const size_t target =  atoi(argv[1]);
  
-  const leaf_t val = 0x12;//_mm_set1_epi8(0x12);
+  const leaf_t val =  3333; //_mm256_set1_epi8(0x01);
 
+ 
   auto [dpfkey0, dpfkey1] = dpf_key<leaf_t, node_t, prgkey_t>::gen(prgkey, nitems, target, val);
-  
+  std::cout << "leaves per node: "  << dpfkey0.leaves_per_node << std::endl;
   const size_t depth = dpfkey1.depth(nitems);  
  
-  std::bitset<depth> directions = target;
+  from_PB_ from_P0_[depth], from_P1_[depth], from_PB_other_[depth], from_PB_other_2[depth];
+
+  from_PB_ from_P0_V0_[depth], from_P1_V0_[depth];
+
+  std::cout << "depth = " << depth << std::endl;
+  std::bitset<depth> directions = ceil(target/dpfkey0.leaves_per_node);
   
   for(std::size_t i = 0; i < depth/2; ++i) 
   {
     bool t = directions[i];
     directions[i] = directions[depth-i-1];
     directions[depth-i-1] = t;
-  }
+  } 
   
+  std::bitset<depth>  P0direction = rand();
+  std::bitset<depth>  P1direction =  P0direction ^ directions;
   std::cout << "directions = " << directions << std::endl;
-  
+   
   AES_KEY aeskey;
   
   from_P2 from_P2_to_P0(depth, rounds), from_P2_to_P1(depth, rounds);
-  from_PB from_P0(depth, rounds), from_P1(depth, rounds);
+  //from_PB from_P0(depth, rounds), from_P1(depth, rounds);
 
   from_P2 from_P2_to_PB(depth, rounds);
-  from_PB from_PB_other(depth, rounds);
+ // from_PB from_PB_other(depth, rounds);
   
+
+  from_P2 from_P2_to_P0_V0(depth, rounds), from_P2_to_P1_V0(depth, rounds);
+ // from_PB from_P0_V0(depth, rounds), from_P1_V0(depth, rounds);
+
   size_t len = 100000;
   
   __m128i seed0, seed1, seed2;
+
   
+//for(size_t ii = 0; ii < number_of_simulations; ++ii)
+{  
+//  std::cout << "ii = " << ii << std::endl;
   arc4random_buf(&seed0, sizeof(__m128i));
   arc4random_buf(&seed1, sizeof(__m128i));
   arc4random_buf(&seed2, sizeof(__m128i));
-  
 
   Simulator sim(aeskey, seed0, seed1, seed2, len, depth);  
+  for(size_t i = 0; i < depth; ++i)
+  {
+    sim.P0direction[i] = P0direction[i];
+    sim.P1direction[i] = P1direction[i];
+  }
+
+  sim.root_layer(from_P0_, from_P1_, from_P2_to_P0, from_P2_to_P1, prgkey, dpfkey0, dpfkey1);
+
+  for(size_t index = 1; index < depth; ++index)
+  {
+   sim.middle_layers(from_P0_, from_P1_, from_P2_to_P0, from_P2_to_P1, prgkey, dpfkey0, dpfkey1, index);
+  }  
+
+  std::cout << "--------------------------------------" << std::endl << std::endl << std::endl;
+}
+
+
+  Verifier2 ver2_0(aeskey, seed0, seed1, seed2, len, depth);  
+
+  for(size_t i = 0; i < depth; ++i)
+  {
+    ver2_0.P0direction[i] = P0direction[i];
+    ver2_0.P1direction[i] = P1direction[i];
+  }
+
+  ver2_0.root_layer(from_P0_V0_, from_P1_V0_, from_P2_to_P0_V0, from_P2_to_P1_V0, prgkey, dpfkey0, dpfkey1);
+ 
+  for(size_t index = 1; index < depth; ++index)
+  {
+   ver2_0.middle_layers(from_P0_V0_, from_P1_V0_, from_P2_to_P0_V0, from_P2_to_P1_V0, prgkey, dpfkey0, dpfkey1, index);
+  }
+  
+
   Verifier  ver0(aeskey, seed0, len, depth);
   Verifier  ver1(aeskey, seed1, len, depth);
-
-  sim.root_layer(from_P0, from_P1, from_P2_to_P0, from_P2_to_P1, prgkey, dpfkey0, dpfkey1);
-
-  std::cout << "dept = " << depth << std::endl;
-
-  for(size_t index = 1; index < depth-1; ++index)
-  {
-   sim.middle_layers(from_P0, from_P1, from_P2_to_P0, from_P2_to_P1, prgkey, dpfkey0, dpfkey1, index);
-  }
-  
-  ver0.Pdirection = sim.P0direction;
-  
-  ver0.root_layer(from_P2_to_P0, from_P1, from_PB_other, prgkey, dpfkey0, party0);
-  
-  assert(from_PB_other.L_shares_recv == from_P0.L_shares_recv);
-  assert(from_PB_other.R_shares_recv == from_P0.R_shares_recv);
-  assert(from_PB_other.bit_L_shares_recv == from_P0.bit_L_shares_recv);
-  assert(from_PB_other.bit_R_shares_recv == from_P0.bit_R_shares_recv);
-  
-  for(size_t j = 0; j < 4; ++j)
-  {
-   assert(from_PB_other.next_bit_L_recv[0][j] == from_P0.next_bit_L_recv[0][j]);
-   assert(from_PB_other.next_bit_R_recv[0][j] == from_P0.next_bit_R_recv[0][j]);
-   assert(from_PB_other.blinds_recv[0][j] == from_P0.blinds_recv[0][j]);   
-   assert(from_PB_other.next_bit_L_recv2[0][j] == from_P0.next_bit_L_recv2[0][j]);
-   assert(from_PB_other.next_bit_R_recv2[0][j] == from_P0.next_bit_R_recv2[0][j]);
-  }
-
-   for(size_t index = 1; index < depth-1; ++index)
-   { 
-       ver0.middle_layers(from_P2_to_P0, from_P1, from_PB_other, prgkey, ver0.seed0[index], ver0.seed1[index],   dpfkey0, index, party0);
-   }
-
-  for(size_t i = 1; i < depth-1; ++i)
-  {
-    for(size_t j = 0; j < 4; ++j)
-    {
-     assert(from_PB_other.next_bit_L_recv[i][j] == from_P0.next_bit_L_recv[i][j]);
-     assert(from_PB_other.next_bit_R_recv[i][j] == from_P0.next_bit_R_recv[i][j]);
-     assert(from_PB_other.blinds_recv[i][j] == from_P0.blinds_recv[i][j]);   
-     assert(from_PB_other.next_bit_L_recv2[i][j] == from_P0.next_bit_L_recv2[i][j]);
-     assert(from_PB_other.next_bit_R_recv2[i][j] == from_P0.next_bit_R_recv2[i][j]);
-
-    }
-
-    for(size_t j = 0; j < rounds; ++j)
-    {
-      assert(from_PB_other.seed0R_encrypt[i][j] == from_P0.seed0R_encrypt[i][j]);
-    } 
-   }
-
-  std::string src = from_PB_other.to_string();
-  std::string src2 = from_P0.to_string();
-  std::cout << "hash: (verifier 0) " << picosha2::hash256_hex_string(src) << "\n" << std::endl;
-  std::cout << "hash: (verifier 0) " << picosha2::hash256_hex_string(src2) << "\n" << std::endl;
-
-  ver1.Pdirection = sim.P1direction;
-
-  std::cout << std::endl << std::endl << std::endl << std::endl << std::endl << std::endl << " ------------------------------------------------------------ " << std::endl << std::endl;
-
-  ver1.root_layer(from_P2_to_P1, from_P0, from_PB_other,   prgkey, dpfkey1, party1);
-
-  assert(from_PB_other.L_shares_recv == from_P1.L_shares_recv);
-  assert(from_PB_other.R_shares_recv == from_P1.R_shares_recv);
-  assert(from_PB_other.bit_L_shares_recv == from_P1.bit_L_shares_recv);
-  assert(from_PB_other.bit_R_shares_recv == from_P1.bit_R_shares_recv);  
  
-  for(size_t j = 0; j < 4; ++j)
-  {
-   assert(from_PB_other.next_bit_L_recv[0][j] == from_P1.next_bit_L_recv[0][j]);
-   assert(from_PB_other.next_bit_R_recv[0][j] == from_P1.next_bit_R_recv[0][j]);
-   assert(from_PB_other.blinds_recv[0][j] == from_P1.blinds_recv[0][j]);   
-   assert(from_PB_other.next_bit_L_recv2[0][j] == from_P1.next_bit_L_recv2[0][j]);
-   assert(from_PB_other.next_bit_R_recv2[0][j] == from_P1.next_bit_R_recv2[0][j]);
+ 
+
+  for(size_t i = 0; i < depth; ++i) ver0.Pdirection[i] = ver2_0.P0direction[i];
+  
+  ver0.root_layer(from_P2_to_P0, from_P1_, from_PB_other_, prgkey, dpfkey0, party0);
+  
+  assert(from_PB_other_[0].L_shares_recv == from_P0_[0].L_shares_recv);
+  assert(from_PB_other_[0].R_shares_recv == from_P0_[0].R_shares_recv);
+  assert(from_PB_other_[0].bit_L_shares_recv == from_P0_[0].bit_L_shares_recv);
+  assert(from_PB_other_[0].bit_R_shares_recv == from_P0_[0].bit_R_shares_recv);
+  
+  for(size_t index = 1; index < depth; ++index)
+  { 
+       ver0.middle_layers(from_P2_to_P0 , from_P1_, from_PB_other_, prgkey, ver0.seed0[index], ver0.seed1[index],   dpfkey0, index, party0);
   }
-  for(size_t index = 1; index < depth-1; ++index)
+  
+  for(size_t i = 0; i < depth-1; ++i)
+  {
+      //std::cout << "i = " << i << std::endl;
+      for(size_t j = 0; j < 4; ++j)
+      { 
+       //std::cout << "-> " << (from_PB_other_[i].blinds_recv[j] ^ from_P0_[i].blinds_recv[j]).bits << std::endl;
+       assert(from_PB_other_[i].next_bit_L_recv[j] == from_P0_[i].next_bit_L_recv[j]);
+       assert(from_PB_other_[i].next_bit_R_recv[j] == from_P0_[i].next_bit_R_recv[j]);
+       assert(from_PB_other_[i].blinds_recv[j] == from_P0_[i].blinds_recv[j]);   
+       //assert(from_PB_other_[i].next_bit_L_recv2[j] == from_P0_[i].next_bit_L_recv2[j]);
+       //assert(from_PB_other_[i].next_bit_R_recv2[j] == from_P0_[i].next_bit_R_recv2[j]);
+      }
+
+      for(size_t j = 0; j < rounds; ++j)
+      {
+        assert(from_PB_other_[i].seed0L_encrypt[j] == from_P0_[i].seed0L_encrypt[j]);
+        assert(from_PB_other_[i].seed0R_encrypt[j] == from_P0_[i].seed0R_encrypt[j]);
+      } 
+  }
+
+
+
+
+  const char * buffer0 = (const  char*)& from_P0_[2];
+  const char * buffer_other_0 = (const  char*)&from_PB_other_[2];
+  std::string str_other_0(buffer_other_0);
+  std::string str0(buffer0);
+  std::cout << "str_other_0 = " << picosha2::hash256_hex_string(str_other_0) << std::endl;
+  std::cout << "str0        = " << picosha2::hash256_hex_string(str0) << std::endl;
+  
+   ver1.Pdirection = ver2_0.P1direction;
+
+   std::cout << std::endl << std::endl << std::endl << std::endl << std::endl << std::endl << " ------------------------------------------------------------ " << std::endl << std::endl;
+
+   ver1.root_layer(from_P2_to_P1 , from_P0_ , from_PB_other_2,   prgkey, dpfkey1, party1);
+
+ 
+ 
+ 
+
+
+  for(size_t index = 1; index < depth; ++index)
   {  
-    ver1.middle_layers(from_P2_to_P1, from_P0, from_PB_other, prgkey, ver1.seed0[index], ver1.seed1[index], dpfkey1, index, party1);
+    ver1.middle_layers(from_P2_to_P1 , from_P0_ , from_PB_other_2 , prgkey, ver1.seed0[index], ver1.seed1[index], dpfkey1, index, party1);
   }
 
  
-  for(size_t i = 1; i < depth-1; ++i)
+  for(size_t i = 0; i < depth-1; ++i)
   {
     for(size_t j = 0; j < 4; ++j)
-    {
-     assert(from_PB_other.next_bit_L_recv[i][j] == from_P1.next_bit_L_recv[i][j]);
-     assert(from_PB_other.next_bit_R_recv[i][j] == from_P1.next_bit_R_recv[i][j]);
-     assert(from_PB_other.blinds_recv[i][j] == from_P1.blinds_recv[i][j]);   
-     assert(from_PB_other.next_bit_L_recv2[i][j] == from_P1.next_bit_L_recv2[i][j]);
-    } 
+    { 
+      assert(from_PB_other_2[i].next_bit_L_recv[j] == from_P1_[i].next_bit_L_recv[j]);
+      assert(from_PB_other_2[i].next_bit_R_recv[j] == from_P1_[i].next_bit_R_recv[j]);
+      assert(from_PB_other_2[i].blinds_recv[j] == from_P1_[i].blinds_recv[j]);   
+      assert(from_PB_other_2[i].next_bit_L_recv2[j] == from_P1_[i].next_bit_L_recv2[j]);
+      assert(from_PB_other_2[i].next_bit_R_recv2[j] == from_P1_[i].next_bit_R_recv2[j]);
+    }
 
     for(size_t j = 0; j < rounds; ++j)
     {
-      assert(from_PB_other.seed1R_encrypt[i][j] == from_P1.seed1R_encrypt[i][j]);
-    }
-
+      assert(from_PB_other_2[i].seed0L_encrypt[j] == from_P1_[i].seed0L_encrypt[j]);
+      assert(from_PB_other_2[i].seed0R_encrypt[j] == from_P1_[i].seed0R_encrypt[j]);
+    } 
   }
- src = from_PB_other.to_string();
- src2 = from_P1.to_string();
-std::cout << "hash: (verifier 1) " << picosha2::hash256_hex_string(src) << "\n" << std::endl;
-std::cout << "hash: (verifier 1) " << picosha2::hash256_hex_string(src2) << "\n" << std::endl;
+ 
+ 
+ 
+  const char * buffer_other = (const  char*)&from_PB_other_2[2];
+
+  const char * buffer1 = (const  char*)&from_P1_[2];
+ 
+ 
+ 
+  std::string str_other(buffer_other);
+  std::string str1(buffer1);    
+ 
+  std::cout << "check0 = " << picosha2::hash256_hex_string(str_other) << std::endl;
+  std::cout << "check1 = " << picosha2::hash256_hex_string(str1) << std::endl;
+  
+ 
+ 
+  char * transcript = reinterpret_cast<char* >(from_PB_other_2);
+  unsigned char hashed[picosha2::k_digest_size];
+  picosha2::hash256(transcript, transcript + depth * sizeof(from_PB_other_2[0]), hashed, hashed + picosha2::k_digest_size);
+ 
+  char * transcript1 = reinterpret_cast<char* >(from_P1_);
+  unsigned char hashed1[picosha2::k_digest_size];
+  picosha2::hash256(transcript1, transcript1 + depth * sizeof(from_P1_[0]), hashed1, hashed1 + picosha2::k_digest_size);
+
+ 
+  for (int i = 0; i < 32; i++) {
+  printf("%x", hashed[i]);
+}
+ std::cout << std::endl << "-- " << std::endl;
+  for (int i = 0; i < 32; i++) {
+  printf("%x", hashed1[i]);
+}
+ // std::cout << "hash: (verifier 1) " << picosha2::hash256_hex_string(src) << "\n" << std::endl;
+ // std::cout << "hash: (verifier 1) " << picosha2::hash256_hex_string(src2) << "\n" << std::endl;
+ 
+ 
+
 
   return 0;
 }
