+++ prg.h
@@ -20,7 +20,8 @@ namespace dpf
 
 template<typename node_t, typename prgkey_t>
 inline void PRG(const prgkey_t & prgkey, const node_t seed, void * outbuf, const uint32_t len, const uint32_t from = 0);
-
+template<typename node_t, typename prgkey_t>
+inline void PRG2(const prgkey_t & prgkey, const node_t seed, void * outbuf, const uint32_t len, const uint32_t from = 0);
 template<>
 inline void PRG(const AES_KEY & prgkey, const __m128i seed, void * outbuf, const uint32_t len, const uint32_t from)
 {
@@ -30,7 +31,7 @@ inline void PRG(const AES_KEY & prgkey, const __m128i seed, void * outbuf, const
 		outbuf128[i] = _mm_xor_si128(seed, _mm_set_epi64x(0, from+i));
 	}
 	AES_ecb_encrypt_blks(outbuf128, static_cast<unsigned int>(len), &prgkey);
-	for (size_t i = 0; i < len; ++i)
+	for (size_t i = 0; i < len; ++i) 
 	{
 		outbuf128[i] = _mm_xor_si128(outbuf128[i], _mm_set_epi64x(0, from+i));
 		outbuf128[i] = _mm_xor_si128(outbuf128[i], seed);
@@ -38,17 +39,35 @@ inline void PRG(const AES_KEY & prgkey, const __m128i seed, void * outbuf, const
 } // PRG
 
 template<>
-inline void PRG(const LowMC<__m256i> & prgkey, const __m256i seed, void * outbuf, const uint32_t len, const uint32_t from)
+inline void PRG2(const LowMC<__m256i> & prgkey, const __m256i seed, void * outbuf, const uint32_t len, const uint32_t from)
 {
+ 
 	__m256i * outbuf256 = reinterpret_cast<__m256i *>(outbuf);
+
 	for (size_t i = 0; i < len; ++i)
 	{
 		auto tmp = _mm256_xor_si256(seed, _mm256_set_epi64x(0, 0, 0, from+i));
+		std::cout << "tmp: " << i << " : (PRG): " << tmp[0] << " " << tmp[1] << " " << tmp[2] << " " << tmp[3] << std::endl;
 		outbuf256[i] = prgkey.encrypt(tmp);
 		outbuf256[i] = _mm256_xor_si256(outbuf256[i], tmp);
 	}
-} // PRG
 
+} // PRG
+template<>
+inline void PRG(const LowMC<__m256i> & prgkey, const __m256i seed, void * outbuf, const uint32_t len, const uint32_t from)
+{
+ 
+	__m256i * outbuf256 = reinterpret_cast<__m256i *>(outbuf);
+ 
+	for (size_t i = 0; i < len; ++i)
+	{
+		auto tmp = _mm256_xor_si256(seed, _mm256_set_epi64x(0, 0, 0, from+i));
+		 
+		outbuf256[i] = prgkey.encrypt(tmp);
+		outbuf256[i] = _mm256_xor_si256(outbuf256[i], tmp);
+	}
+
+} // PRG
 template<>
 inline void PRG(const LowMC<__m128i> & prgkey, const __m128i seed, void * outbuf, const uint32_t len, const uint32_t from)
 {
